Version 1.2 20080520
The “sensorless” version. For all of you who don’t want to use a
(motorized) cover with sensors or switches this is the version to use.
Set the variable sensorless on line 33 in the file [CD-Pro2M_main.asm]
to “1″ to build the sensorless version. The open/close buttons on the
front panel and on the remote control unit do not have any function in
this sensorless version.